[nrf fromlist] soc: nrf54h: gpd: use callback to fetch nrfs async result
Busy-waiting for the result of the nrfs service calls can stall, so
let's use a callback that flags a semaphore instead. Since the API is
supposed to be callable in the context of pre-kernel, fallback to
busy-wait on that scenario.

Upstream PR #: 80819

Signed-off-by: Gerard Marull-Paretas <gerard@teslabs.com>
(cherry picked from commit 714d82d232977309e0bdc8dadd2d7a71ee914a4b)