[nrf noup] drivers: hwinfo: Support for reset reasons in nRF54H20
Adding support for reset reasons in the nRF54H20 SoC.

Upstream PR #: 81751

This commit should be replaced with a fromtree one, since PR 81751 has
been already merged, but the latter became dependent on nrfx 3.10.0
which is not present in the upmerge and integrating it at the final
stage of the upmerge is too risky, hence the commit has been turned
into a noup. It is supposed to be replaced with the proper fromtree
commit right after the upmerge.

Signed-off-by: Karol Lasończyk <karol.lasonczyk@nordicsemi.no>