[nrf noup] samples: drivers: jesd216: Remove unnecessary Kconfig option changes
It is no longer needed to enable a specific flash driver in the sample
configuration. An appropriate driver (together with its dependencies,
like the SPI bus) for a given target will be enabled automatically
based on what is enabled in DTS.

Upstream PR #: 80042
Applied as a [nrf noup] due to a conflict.

Signed-off-by: Andrzej Głąbek <andrzej.glabek@nordicsemi.no>