[nrf fromlist] drivers: clock_control_nrf2: Add missing cancelation of request
This is a follow-up to commit fe0e2dbc6044ffa74b27bfcba335980bc5a6ae64.

If `nrf_clock_control_request_sync()` ends up with a timeout, before
returning it must cancel the request that was not fulfilled on time.
Otherwise, the request may actually finish succesfully a bit later,
but the caller will not be aware that the clock needs to be released
(since the call resulted in an error).

Upstream PR #: 82994

Signed-off-by: Andrzej Głąbek <andrzej.glabek@nordicsemi.no>