boards/mediatek: Fix DTS mistakes
The two interrupt controller registers on 818x got spoonerized, oops.
Hilariously it actually worked (mostly) becuase the interrupt would
still be delivered, the controller isn't involved in the ACK/clear
process on thsi device, the and the set of pending interrupts was
masked against enabled ones.  Found it by accident when I moved things
around and caught a "spurious" interrupt.

Also the 8195 DRAM region was all packed in a single block that doesn't
honor the two regions defined in the host devicetree.  That's benign
from Zephyr (the memory in the hole actually is usable) but dangerous
for SOF as the second region is used by the host driver to allocate
DMA buffers and we can't touch it except as directed.  Honor the
scheme used by the other devices.

Signed-off-by: Andy Ross <andyross@google.com>