boards: sparkfun: pro_micro_rp2040: fix pinctrl definitions for spi
The default SCK and COPI/MOSI pins on the Sparkfun Pro Micro RP2040
board correspond to `GPIO22` and `GPIO23` respectively[^1].
The existing pinctrl definitions in the `spi0_default` don't reflect
this (rather, they match the pinout of the Adafruit KB2040 MCU[^2]).

Update the board's respective pinctrl definitions to match the pinout.

[^1]: https://cdn.sparkfun.com/assets/e/2/7/6/b/ProMicroRP2040_Graphical_Datasheet.pdf
[^2]: https://learn.adafruit.com/assets/106984

Signed-off-by: Anant Thazhemadam <anant.thazhemadam@gmail.com>