[nrf noup] drivers: clock_control: nrf: skip nrf_clock.h on haltium
Don't include hal/nrf_clock.h on Haltium platforms as that HAL currently
does not compile for those. Since this issue only affects
projects that are downstream of NCS, and should be fixed soon by
NRFX-6892, this patch is made as a temporary no-up that can be reverted
along with the introduction of the fix.

Ref: NCSDK-31213

Signed-off-by: Jonathan Nilsen <jonathan.nilsen@nordicsemi.no>