[nrf fromtree] Bluetooth: Host: Specify parameter needs of `bt_set_name`
State that the name given to `bt_set_name` must be null terminated.

If not, the behavior of the function may be wrong as it's using `strlen`
to get the length of the given name and `strcmp` to compare it to the
current device name. Both of those functions expect null terminated string.

Signed-off-by: Théo Battrel <theo.battrel@nordicsemi.no>
(cherry picked from commit 7884b929fe7635dcde8f68bb60ae2018a572f1fe)
Signed-off-by: alperen sener <alperen.sener@nordicsemi.no>