[nrf fromtree] doc boards nrfbsim: Update list of supported HW
The models include enough of the CRACEN to run the same
nrfx drivers as in Zephyr. So let's add it to the list.

Also let's remove the RTC from the list as the hal does not
expose it anymore for this platform as the GRTC is to be used
instead.

Signed-off-by: Alberto Escolar Piedras <alberto.escolar.piedras@nordicsemi.no>
(cherry picked from commit a211999141833bba7dfd3336f8e3b9b135bb4627)
(cherry picked from commit 5d28a5eb210b8c59eee2b81f37ae939838f91c2f)