[nrf fromtree] boards: nrf*_bsim: Detect attempt to configure not existing int
Similar to 923d4fbad8. Prevent overrunning the irq vector table.
This is not happening today in tree, but coverity thinks it
does. Checking for it to prevent it is not a bad idea
anyhow, so let's do it.

Signed-off-by: Krzysztof Chruściński <krzysztof.chruscinski@nordicsemi.no>
(cherry picked from commit cd23fdf63fd3d15f38f29686fc29454b1d07a2ee)