[nrf fromtree] boards: nrf_bsim: Document how to use the console with bsim
This commit describes how to build and run applications using
the console. This documentation is added because it may
not be obvious how this is done. That is, sometimes a user
may believe that the terminal launching the application is the
the console which is not the case.

Signed-off-by: Rubin Gerritsen <rubin.gerritsen@nordicsemi.no>
(cherry picked from commit 0f3e6fd3d624f3647005b8f2bc30bd477933a60a)
(cherry picked from commit 63aa79c125728a7c89d58f27fc9367214db0545f)