[nrf noup] tests: secure_storage: fix test w/ ZMS backend on 54L15
noup because it's about partition manager.

Fix the build of secure_storage.psa.its.secure_storage.store.zms on
nrf54l15dk/nrf54l15/cpuapp by disabling partition manager, which is
incompatible with the ZMS implementation of the ITS store module.

Disabling it only for that test as it's not needed for the others
and even makes the NS board targets fail if disabling PM.

Signed-off-by: Tomi Fontanilles <tomi.fontanilles@nordicsemi.no>