[nrf noup] bluetooth: Temporary Kconfig fix for BT RPC configuration
The BT_DRIVERS symbol default value 'y' used to depend on !BT_CTLR
but now it is always on when BT is set. For BT_RPC the BT_DRIVERS
symbol must not be enabled on the client side as no driver is used.

The temporary solution is to set BT_DRIVERS to 'y' by default only when
BT_HCI stack selection is enabled. It will be 'n' when BT_RPC_STACK
is enabled. The fix should be fine as NCS uses either HCI or RPC stack.

Signed-off-by: Michał Grochala <michal.grochala@nordicsemi.no>