sysbuild: Autogenerate secondary slot
Generation of the secendary slot is not trivial if one considers all of
the configuration options available in sysbuild, such as:
 - per-image overrides through variables
 - setting FILE_SUFFIX
 - additional images appearing as a result of enabling SNIPPETs
Achieving this through the regular approach (adding images), requires
enumerating all slotted images in CMake files, applying the same set
of overlays manually, which partially makes the configurability
provided by sysbuild nearly unusable for the secondary slot.

Since sysbuild is an entity that is aware of the images being
built and their configuration parameters, this commit tries to rerun the
image build if the user provides the <image>_slot_b.overlay.
There is no need to add Kconfig overlays as the main point of slotted
images is that exactly two identical configurations are built, with only
one difference - the start address, controlled through the
zephyr,code-partition symbol.

Ref: NCSDK-NONE

Signed-off-by: Tomasz Chyrowicz <tomasz.chyrowicz@nordicsemi.no>