[nrf fromlist] dts: nordic: nrf5340: Revert nRF5340 IPC backend to rpmsg
This reverts commit 70419bdee7ffd4ea7248f2b3f4c1a5a9d5a75c76.

This is because there are issues around slow IPC thoughput
with icbmsg, which is causing issues with BLE when lots of
data is required to be exchanged, e.g. with ISO.

Also there is an assert icmsg.c#L190 which occurs when
initializing bluetooth and IPC in certain circumstances.

Upstream PR #: 85621

Signed-off-by: Sean Madigan <sean.madigan@nordicsemi.no>
(cherry picked from commit baaad1a97f8792d00cd264d81b0dbc5a089d5330)