[nrf noup] lib: net_buf: buf: Revert alloc DBG to WRN change
nrf noup resolution - Please drop this commit after upmerge
if there is a conflict, as it should no longer be required.

This warning is raised in many places when using bluetooth.
However, it is not very useful for the user to see.

This is a noup PR as there is currently more changes coming
in around this code and how to handle deadlocks in zephyr.

Signed-off-by: Sean Madigan <sean.madigan@nordicsemi.no>