[nrf noup] soc: nrf54l: Non-XIP KMU reserved memory fix
nrf-squash! [nrf noup] soc: nrf54l: Add custom section for KMU

The linker script inclusion which places the KMU reserved
buffer on the top of RAM doesn't work for non-XIP builds.
The Zephyr linker script will firstly load the code for
an non-XIP build in RAM and then include this KMU related
linker script which results in an unpredictable placement
of the KMU reserved area and a failed build.

In order to support non-XIP builds the linker file is not
included and the a DTS reserved-memory entry should be used.

To limit the scope, the DTS reserved memory region
is currently only supported for non-XIP builds.

This is a noup since the KMU is not supported upstream.

Signed-off-by: Georgios Vasilakis <georgios.vasilakis@nordicsemi.no>