[nrf fromlist] tests: drivers: watchdog: Add test for initialization of .bss and .data
In some configurations, code is copied from non-volatile memory
to RAM before a core (using that code) is booted.
When that core is reset by expired watchdog, variables shall be
restored to "initial" values.

This is problematic because the core is aware of copy in RAM.
It may not know address in non-volatile memory, where "original"
code is stored.

Add tests which verifiest that .bss and .data sections are
correct when core boots from reset caused by expired watchdog.

Upstream PR #: 86217

Signed-off-by: Sebastian Głąb <sebastian.glab@nordicsemi.no>