[nrf fromlist] settings: zms: load only subtree in the argument
If the subtree argument is not NULL in the settings_load function, load
only the setting entry that corresponds to that subtree.
If the subtree argument is NULL or it is not found in the storage, load
all the existing entries in the storage.

Upstream PR #: 87792

Signed-off-by: Riadh Ghaddab <rghaddab@baylibre.com>
(cherry picked from commit 86132feb5c59e8a071d9782c5482ab346cfbf6f3)
(cherry picked from commit df20b6124c8896191f418c251b494119adc6b542)