[nrf fromtree] Bluetooth: Mesh: Fix proxy after board reboot
When the board has been provisioned and the board is rebooted
it was not possible to connect to the board through the proxy
servcie. This fixes that the Mesh Proxy servive is restarted
after the board is rebooted.

Signed-off-by: Ingar Kulbrandstad <ingar.kulbrandstad@nordicsemi.no>
(cherry picked from commit db84181c228ecf2054fd79470b00d9cce468f064)
Signed-off-by: Pavel Vasilyev <pavel.vasilyev@nordicsemi.no>