[nrf fromlist] samples: net: wifi: Add two variants of certs
RSA3K based certs are not supported on all platforms, so, keep both
variants, rsa2k (the older certs but with longer expiry 9999 days) and
rsa3k (latest ones) and we can have more variants in this folders.

Also, add a cmake variable to override the path with default as rsa3k.

Upstream PR #: 87656

Signed-off-by: Chaitanya Tata <Chaitanya.Tata@nordicsemi.no>