[nrf fromlist] samples: drivers: Add spis_wakeup sample
Sample showing how to use the additional wakeup signal
in nordic,nrf-spis driver. The application, configured
as a SPIS, is put to sleep while waiting for an SPI
transmission. An external device working as a SPIM and
also handling the wakeup signal wakes the application
up just before the transmission starts.

Upstream PR #: 87968

Signed-off-by: Piotr Krzyzanowski <piotr.krzyzanowski@nordicsemi.no>